+++ tests/test_stat_count_ownership_or_control_statement_with_at_least_one_interest_beneficial.py
@@ -4,22 +4,28 @@ import tempfile
 from libcovebods.api import bods_json_output
 
 
-def test_not_in_old_schema():
+def test_schema_0_2():
 
     cove_temp_folder = tempfile.mkdtemp(
         prefix="lib-cove-bods-tests-", dir=tempfile.gettempdir()
     )
     json_filename = os.path.join(
-        os.path.dirname(os.path.realpath(__file__)), "fixtures", "0.1", "basic_1.json"
+        os.path.dirname(os.path.realpath(__file__)), "fixtures", "0.2", "basic_1.json"
     )
 
     results = bods_json_output(cove_temp_folder, json_filename)
 
-    assert results["schema_version"] == "0.1"
-    assert "count_person_statements_have_pep_status" not in results["statistics"]
+    assert results["schema_version"] == "0.2"
+
+    assert (
+        1
+        == results["statistics"][
+            "count_ownership_or_control_statement_with_at_least_one_interest_beneficial"
+        ]
+    )
 
 
-def test_1():
+def test_schema_0_2_statement_is_component_but_is_not_used_1():
 
     cove_temp_folder = tempfile.mkdtemp(
         prefix="lib-cove-bods-tests-", dir=tempfile.gettempdir()
@@ -28,40 +34,37 @@ def test_1():
         os.path.dirname(os.path.realpath(__file__)),
         "fixtures",
         "0.2",
-        "pep_status_1.json",
+        "statement_is_component_but_is_not_used_1.json",
     )
 
     results = bods_json_output(cove_temp_folder, json_filename)
 
     assert results["schema_version"] == "0.2"
-    assert results["statistics"]["count_person_statements_have_pep_status"] == 1
+
     assert (
-        results["statistics"][
-            "count_person_statements_have_pep_status_and_reason_missing_info"
+        2
+        == results["statistics"][
+            "count_ownership_or_control_statement_with_at_least_one_interest_beneficial"
         ]
-        == 0
     )
 
 
-def test_missing_1():
+def test_schema_0_3():
 
     cove_temp_folder = tempfile.mkdtemp(
         prefix="lib-cove-bods-tests-", dir=tempfile.gettempdir()
     )
     json_filename = os.path.join(
-        os.path.dirname(os.path.realpath(__file__)),
-        "fixtures",
-        "0.2",
-        "pep_status_missing_info_1.json",
+        os.path.dirname(os.path.realpath(__file__)), "fixtures", "0.3", "basic_1.json"
     )
 
     results = bods_json_output(cove_temp_folder, json_filename)
 
-    assert results["schema_version"] == "0.2"
-    assert results["statistics"]["count_person_statements_have_pep_status"] == 1
+    assert results["schema_version"] == "0.3"
+
     assert (
-        results["statistics"][
-            "count_person_statements_have_pep_status_and_reason_missing_info"
+        1
+        == results["statistics"][
+            "count_ownership_or_control_statement_with_at_least_one_interest_beneficial"
         ]
-        == 1
     )